[2067] Agreement negotiation: Part 2.3
- Move permissions to dedicated permissions.py file
