+++ vitrina/smart_contracts/services.py
@@ -25,9 +25,7 @@ from vitrina.smart_contracts.utils import (
 )
 from vitrina.users.models import User
 from vitrina.smart_contracts.models import Agreement, AgreementStatuses
-from vitrina.projects.models import Project
 from django.db.models import Q, QuerySet
-from django.contrib.auth.models import AnonymousUser
 from django.utils.translation import gettext_lazy as _
 
 logger = logging.getLogger()
@@ -216,7 +214,7 @@ def extract_elements_from_adoc(adoc_path: str, regex: str) -> list[str]:
             os.remove(TEMP_PDF_PATH)
 
 
-def get_agreements(user: User) -> QuerySet["Agreement"]:
+def get_agreements(user: User) -> QuerySet[Agreement]:
     if not user.is_authenticated:
         return Agreement.objects.none()
 
@@ -227,35 +225,3 @@ def get_agreements(user: User) -> QuerySet["Agreement"]:
     queryset = Agreement.objects.filter(Q(assignee_id__in=represented_org_ids) | Q(assigner_id__in=represented_org_ids))
 
     return queryset
-
-
-def can_view_agreements(user: User | AnonymousUser, project: Project) -> bool:
-    if not user.is_authenticated:
-        return False
-
-    represented_org_ids = user.represented_org_ids
-
-    if user.is_staff or user.is_superuser:
-        return True
-
-    if project.organization and project.organization.id in represented_org_ids:
-        return True
-
-    return project.agreements.filter(assigner_id__in=represented_org_ids).exists()
-
-
-def can_view_agreement(user: User, agreement: Agreement) -> bool:
-    return get_agreements(user).filter(pk=agreement.pk).exists()
-
-
-def can_create_agreements(user: User, project: Project) -> bool:
-    if project.organization:
-        return project.organization == user.viisp_organization and user.is_representative_of(project.organization, True)
-
-    return False
-
-
-def can_upload_agreement_file(user: User, agreement: Agreement) -> bool:
-    parties = [agreement.assignee, agreement.assigner]
-
-    return any(user.viisp_organization == party and user.is_representative_of(party, True) for party in parties)
